requirements.txt
tweepy==4.1.0
python-dotenv 0.19.1